Move length pipe logic into formatLength().
