+++ src/app/shared/formatters/digits-info.ts
