+++ src/app/shared/formatters/format-length.ts
@@ -1,0 +1,28 @@
+import { formatNumber } from "@angular/common";
+import { MeasurementSystem } from "../services/settings/measurement-system";
+import { DigitsInfo } from "./digits-info";
+
+function convertMetersToCustomary(value: number) {
+  return value / 0.3408;
+}
+
+export function formatLength(
+  measurementSystem: MeasurementSystem,
+  localeId: string,
+  valueInMeters: number,
+  digitsInfo?: DigitsInfo,
+) {
+  const value =
+    measurementSystem === MeasurementSystem.metric
+      ? valueInMeters
+      : convertMetersToCustomary(valueInMeters);
+
+  const formatted = formatNumber(value, localeId, digitsInfo);
+
+  const suffix =
+    measurementSystem === MeasurementSystem.metric
+      ? 'm'
+      : 'ft';
+
+  return `${formatted} ${suffix}`;
+}
